fixing Makefile

+++ Makefile
@@ -9,18 +9,14 @@ install: query deauth auth database.db
 	[ -d ${DESTDIR}/db ] ||  mkdir -p ${DESTDIR}/db
 	chown www:www ${DESTDIR}/db
 	chmod 0700 ${DESTDIR}/db
-ifeq($(REPLACEDB),true)
-	install -o www -g www -m 0600 database.db ${DESTDIR}/db
-endif
+	test [ "${REPLACEDB}" = "true" ] && install -o www -g www -m 0600 database.db ${DESTDIR}/db
 	install -o www -g www -m 0500 query ${DESTDIR}/query
 	install -o www -g www -m 0500 auth ${DESTDIR}/auth
 	install -o www -g www -m 0500 deauth ${DESTDIR}/deauth
 
 install-pubnix: query deauth auth database.db
 	[ -d ${PBNIX_HTML}/db ] ||  mkdir -p ${PBNIX_HTML}/db
-ifeq(${REPLACEDB},true)
-	install -m 0666 database.db ${PBNIX_HTML}/db
-endif
+	test [ "${REPLACEDB}" = "true" ] && install -m 0666 database.db ${PBNIX_HTML}/db
 	install -m 0755 query ${PBNIX_HTML}/query.cgi
 	install -m 0755 auth ${PBNIX_HTML}/auth.cgi
 	install -m 0755 deauth ${PBNIX_HTML}/deauth.cgi
